Moving CrumbExclusion implementation to it's own class and some further cleanup
